build: track if we have stdbool.h at configure time.
This adds BSON_HAVE_STDBOOL_H to bson-config.h so that we can easily
include <stdbool.h> when available. The previous check was not ever
getting used by external software because they would also have to
define HAVE_STDBOOL_H, which nobody does. And of course, this causes
inconsistencies in function calls.